refactor: resource 패지키 생성 및 dao -> repository 이름 변경
